requirements.txt
Flask==2.1.3
Werkzeug==2.1.2
loguru==0.7.0
pycryptodome==3.19.0
python-dotenv==1.0.0
gunicorn==20.1.0